requirements.txt
aiogram==2.17.0
apscheduler==3.8.0
sqlalchemy==1.4.23
asyncpg==0.25.0
python-dotenv==0.19.1
envparse==0.2.0
psycopg2==2.9.1
schedule==1.2.1